+++ roktux/src/test/java/com/rokt/roktux/component/ColumnComponentTest.kt
@@ -14,6 +14,7 @@ import com.rokt.core.testutils.annotations.DcuiConfig
 import com.rokt.core.testutils.annotations.DcuiNodeComponentState
 import com.rokt.core.testutils.annotations.DcuiNodeJson
 import com.rokt.core.testutils.annotations.WindowSize
+import com.rokt.core.testutils.assertion.assertAlpha
 import com.rokt.core.testutils.assertion.assertBackgroundColor
 import com.rokt.core.testutils.assertion.assertHeightFit
 import com.rokt.core.testutils.assertion.assertHeightWrapContent
@@ -259,5 +260,12 @@ class ColumnComponentTest : BaseDcuiEspressoTest() {
     fun testRowComponentGapProperty() {
         composeTestRule.onNodeWithTag(DCUI_COMPONENT_TAG).assertIsDisplayed()
             .assertHeightIsEqualTo(80.dp)
+    }
+
+    @Test
+    @DcuiNodeJson(jsonFile = "ColumnComponent/Column_with_Opacity.json")
+    fun testColumnComponentWithOpacity() {
+        composeTestRule.onNodeWithTag(DCUI_COMPONENT_TAG)
+            .assertAlpha(0.7f)
     }
 }